infrastructure/pyproject.toml
[project]
name = "stonksfeed-infrastructure"
version = "0.1.0"
description = "AWS CDK infrastructure for Stonksfeed"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "boto3>=1.35",
    "botocore>=1.35",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
testpaths = ["lambdas"]
python_files = ["test_*.py"]
